prevents scrolling of the body while the lighbox is active
this time, not by trying to stop the scroll event or touch event or
wheel event or whatever to propagate or activate or whatever. But by
hard-fixing the document position in place for the duration of the
lightbox.

Thus now on desktop, you can zoom the lighbox by scrolling on the
image, which zooms the image. On mobile, you can pinch-to-zoom like
always. This was way harder than necessary.